Move checkPermission method from UserController to UserService
