+++ src/main/java/com/codessquad/qna/service/UserService.java
@@ -1,14 +1,16 @@
 package com.codessquad.qna.service;
 
 import com.codessquad.qna.domain.User;
-import com.codessquad.qna.exception.DuplicateUserIdFoundException;
-import com.codessquad.qna.exception.UserNotFoundException;
+import com.codessquad.qna.exception.*;
 import com.codessquad.qna.repository.UserRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
-import sun.util.resources.cldr.aa.CurrencyNames_aa_ER;
 
+import javax.servlet.http.HttpSession;
 import java.util.List;
+
+import static com.codessquad.qna.controller.HttpSessionUtils.getSessionUser;
+import static com.codessquad.qna.controller.HttpSessionUtils.isLoginUser;
 
 
 @Service
@@ -43,6 +45,22 @@ public class UserService {
 
     public User findByUserId(String userId) {
         return userRepository.findByUserId(userId).orElseThrow(UserNotFoundException::new);
+    }
+
+    public User findVerifiedUser(Long id, HttpSession session) {
+        User user = userRepository.findById(id).orElseThrow(UserNotFoundException::new);
+        checkPermission(id, session);
+        return user;
+    }
+
+    private void checkPermission(Long id, HttpSession session) {
+        if (!isLoginUser(session)) {
+            throw new FailedUserLoginException();
+        }
+        User loginUser = getSessionUser(session);
+        if (!loginUser.matchId(id)) {
+            throw new IllegalUserAccessException();
+        }
     }
 }
